move around some code
